+++ stripe/_invoice_line_item_service.py
@@ -53,7 +53,7 @@ class InvoiceLineItemService(StripeService):
         """
         metadata: NotRequired["Literal['']|Dict[str, str]"]
         """
-        Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format. Individual keys can be unset by posting an empty value to them. All keys can be unset by posting an empty value to `metadata`.
+        Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format. Individual keys can be unset by posting an empty value to them. All keys can be unset by posting an empty value to `metadata`. For `type=recurring` line items, the incoming metadata specified on the request is directly used to set this value, in contrast to `type=invoiceitem` line items, where any existing metadata on the invoice line is merged with the incoming data.
         """
         period: NotRequired["InvoiceLineItemService.UpdateParamsPeriod"]
         """