+++ stripe/_login_link.py
@@ -7,7 +7,7 @@ from typing_extensions import Literal
 
 class LoginLink(StripeObject):
     """
-    Login Links are single-use login link for an Express account to access their Stripe dashboard.
+    Login Links are single-use URLs for a connected account to access the Express Dashboard. The connected account's [account.controller.stripe_dashboard.type](https://stripe.com/api/accounts/object#account_object-controller-stripe_dashboard-type) must be `express` to have access to the Express Dashboard.
     """
 
     OBJECT_NAME: ClassVar[Literal["login_link"]] = "login_link"