+++ stripe/_price_service.py
@@ -90,7 +90,7 @@ class PriceService(StripeService):
         """
         unit_amount: NotRequired[int]
         """
-        A positive integer in cents (or local equivalent) (or 0 for a free price) representing how much to charge. One of `unit_amount` or `custom_unit_amount` is required, unless `billing_scheme=tiered`.
+        A positive integer in cents (or local equivalent) (or 0 for a free price) representing how much to charge. One of `unit_amount`, `unit_amount_decimal`, or `custom_unit_amount` is required, unless `billing_scheme=tiered`.
         """
         unit_amount_decimal: NotRequired[str]
         """