+++ stripe/_product.py
@@ -85,7 +85,7 @@ class Product(
     class CreateFeatureParams(RequestOptions):
         entitlement_feature: str
         """
-        The ID of the [Feature](docs/api/entitlements/feature) object attached to this product.
+        The ID of the [Feature](https://stripe.com/docs/api/entitlements/feature) object attached to this product.
         """
         expand: NotRequired[List[str]]
         """