+++ stripe/_setup_attempt.py
@@ -41,6 +41,9 @@ class SetupAttempt(ListableAPIResource["SetupAttempt"]):
         class AcssDebit(StripeObject):
             pass
 
+        class AmazonPay(StripeObject):
+            pass
+
         class AuBecsDebit(StripeObject):
             pass
 
@@ -331,6 +334,9 @@ class SetupAttempt(ListableAPIResource["SetupAttempt"]):
         class Paypal(StripeObject):
             pass
 
+        class RevolutPay(StripeObject):
+            pass
+
         class SepaDebit(StripeObject):
             pass
 
@@ -374,6 +380,7 @@ class SetupAttempt(ListableAPIResource["SetupAttempt"]):
             pass
 
         acss_debit: Optional[AcssDebit]
+        amazon_pay: Optional[AmazonPay]
         au_becs_debit: Optional[AuBecsDebit]
         bacs_debit: Optional[BacsDebit]
         bancontact: Optional[Bancontact]
@@ -385,6 +392,7 @@ class SetupAttempt(ListableAPIResource["SetupAttempt"]):
         klarna: Optional[Klarna]
         link: Optional[Link]
         paypal: Optional[Paypal]
+        revolut_pay: Optional[RevolutPay]
         sepa_debit: Optional[SepaDebit]
         sofort: Optional[Sofort]
         type: str
@@ -394,6 +402,7 @@ class SetupAttempt(ListableAPIResource["SetupAttempt"]):
         us_bank_account: Optional[UsBankAccount]
         _inner_class_types = {
             "acss_debit": AcssDebit,
+            "amazon_pay": AmazonPay,
             "au_becs_debit": AuBecsDebit,
             "bacs_debit": BacsDebit,
             "bancontact": Bancontact,
@@ -405,6 +414,7 @@ class SetupAttempt(ListableAPIResource["SetupAttempt"]):
             "klarna": Klarna,
             "link": Link,
             "paypal": Paypal,
+            "revolut_pay": RevolutPay,
             "sepa_debit": SepaDebit,
             "sofort": Sofort,
             "us_bank_account": UsBankAccount,