+++ stripe/_token.py
@@ -35,7 +35,7 @@ class Token(CreateableAPIResource["Token"]):
 
     You can't store or use tokens more than once. To store card or bank account
     information for later use, create [Customer](https://stripe.com/docs/api#customers)
-    objects or [Custom accounts](https://stripe.com/docs/api#external_accounts).
+    objects or [External accounts](https://stripe.com/api#external_accounts).
     [Radar](https://stripe.com/docs/radar), our integrated solution for automatic fraud protection,
     performs best with integrations that use client-side tokenization.
     """
@@ -1068,11 +1068,12 @@ class Token(CreateableAPIResource["Token"]):
     """
     These bank accounts are payment methods on `Customer` objects.
 
-    On the other hand [External Accounts](https://stripe.com/docs/api#external_accounts) are transfer
-    destinations on `Account` objects for [Custom accounts](https://stripe.com/docs/connect/custom-accounts).
+    On the other hand [External Accounts](https://stripe.com/api#external_accounts) are transfer
+    destinations on `Account` objects for accounts where [controller.requirement_collection](https://stripe.com/api/accounts/object#account_object-controller-requirement_collection)
+    is `application`, which includes [Custom accounts](https://stripe.com/connect/custom-accounts).
     They can be bank accounts or debit cards as well, and are documented in the links above.
 
-    Related guide: [Bank debits and transfers](https://stripe.com/docs/payments/bank-debits-transfers)
+    Related guide: [Bank debits and transfers](https://stripe.com/payments/bank-debits-transfers)
     """
     card: Optional["Card"]
     """
@@ -1115,7 +1116,7 @@ class Token(CreateableAPIResource["Token"]):
     def create(cls, **params: Unpack["Token.CreateParams"]) -> "Token":
         """
         Creates a single-use token that represents a bank account's details.
-        You can use this token with any API method in place of a bank account dictionary. You can only use this token once. To do so, attach it to a [Custom account](https://stripe.com/docs/api#accounts).
+        You can use this token with any API method in place of a bank account dictionary. You can only use this token once. To do so, attach it to a [connected account](https://stripe.com/docs/api#accounts) where [controller.requirement_collection](https://stripe.com/api/accounts/object#account_object-controller-requirement_collection) is application, which includes Custom accounts.
         """
         return cast(
             "Token",
@@ -1132,7 +1133,7 @@ class Token(CreateableAPIResource["Token"]):
     ) -> "Token":
         """
         Creates a single-use token that represents a bank account's details.
-        You can use this token with any API method in place of a bank account dictionary. You can only use this token once. To do so, attach it to a [Custom account](https://stripe.com/docs/api#accounts).
+        You can use this token with any API method in place of a bank account dictionary. You can only use this token once. To do so, attach it to a [connected account](https://stripe.com/docs/api#accounts) where [controller.requirement_collection](https://stripe.com/api/accounts/object#account_object-controller-requirement_collection) is application, which includes Custom accounts.
         """
         return cast(
             "Token",