+++ stripe/identity/_verification_session_service.py
@@ -45,11 +45,11 @@ class VerificationSessionService(StripeService):
         """
         type: NotRequired[Literal["document", "id_number"]]
         """
-        The type of [verification check](https://stripe.com/docs/identity/verification-checks) to be performed.
+        The type of [verification check](https://stripe.com/docs/identity/verification-checks) to be performed. You must provide a `type` if not passing `verification_flow`.
         """
         verification_flow: NotRequired[str]
         """
-        The ID of a Verification Flow from the Dashboard.
+        The ID of a Verification Flow from the Dashboard. See https://docs.stripe.com/identity/verification-flows.
         """
 
     class CreateParamsOptions(TypedDict):
@@ -58,18 +58,6 @@ class VerificationSessionService(StripeService):
         ]
         """
         Options that apply to the [document check](https://stripe.com/docs/identity/verification-checks?type=document).
-        """
-        email: NotRequired[
-            "Literal['']|VerificationSessionService.CreateParamsOptionsEmail"
-        ]
-        """
-        Options that apply to the email check.
-        """
-        phone: NotRequired[
-            "Literal['']|VerificationSessionService.CreateParamsOptionsPhone"
-        ]
-        """
-        Options that apply to the phone check.
         """
 
     class CreateParamsOptionsDocument(TypedDict):
@@ -90,18 +78,6 @@ class VerificationSessionService(StripeService):
         require_matching_selfie: NotRequired[bool]
         """
         Capture a face image and perform a [selfie check](https://stripe.com/docs/identity/verification-checks?type=selfie) comparing a photo ID and a picture of your user's face. [Learn more](https://stripe.com/docs/identity/selfie).
-        """
-
-    class CreateParamsOptionsEmail(TypedDict):
-        require_verification: NotRequired[bool]
-        """
-        Request one time password verification of `provided_details.email`.
-        """
-
-    class CreateParamsOptionsPhone(TypedDict):
-        require_verification: NotRequired[bool]
-        """
-        Request one time password verification of `provided_details.phone`.
         """
 
     class CreateParamsProvidedDetails(TypedDict):
@@ -209,18 +185,6 @@ class VerificationSessionService(StripeService):
         """
         Options that apply to the [document check](https://stripe.com/docs/identity/verification-checks?type=document).
         """
-        email: NotRequired[
-            "Literal['']|VerificationSessionService.UpdateParamsOptionsEmail"
-        ]
-        """
-        Options that apply to the email check.
-        """
-        phone: NotRequired[
-            "Literal['']|VerificationSessionService.UpdateParamsOptionsPhone"
-        ]
-        """
-        Options that apply to the phone check.
-        """
 
     class UpdateParamsOptionsDocument(TypedDict):
         allowed_types: NotRequired[
@@ -240,18 +204,6 @@ class VerificationSessionService(StripeService):
         require_matching_selfie: NotRequired[bool]
         """
         Capture a face image and perform a [selfie check](https://stripe.com/docs/identity/verification-checks?type=selfie) comparing a photo ID and a picture of your user's face. [Learn more](https://stripe.com/docs/identity/selfie).
-        """
-
-    class UpdateParamsOptionsEmail(TypedDict):
-        require_verification: NotRequired[bool]
-        """
-        Request one time password verification of `provided_details.email`.
-        """
-
-    class UpdateParamsOptionsPhone(TypedDict):
-        require_verification: NotRequired[bool]
-        """
-        Request one time password verification of `provided_details.phone`.
         """
 
     class UpdateParamsProvidedDetails(TypedDict):